327-57 Add Java Script functionality
Adds the ability to build and run js through gradle & node. Everything
is handeled by gradle so there is no need to install node or npm etc. In
order to run js scripts, commands must be added to the build.gradle to
do so. It is a very simple process and documenation for it is linked
above the plugin declaration in the build.gradle file.